refactor: api mutation/query 구분
